rewrote outputparser
yep i finally rewrote that piece of crap code. It was using strings as
it was first written in C#. After this i just transferred it to
javascript and then typescript without writing it in a way it *should*
be written. The reason why it before parsed strings is that this made
the communication between js & C# much easier